Created some frankencode which makes the quads face the camera.
Turns out we both worked on the same idea at the same time and got two different solutions. My code managed to rotate the quads over all axes instead of just XZ, so while the main organization of the code is Brendon's, the actual point calculation is my own. It's a strange mix, but it appears to work.